// src/components/Typography/Typography.types.ts
export type TypographyVariant =
  | 'heading1'
  | 'heading2'
  | 'heading3'
  | 'heading4'
  | 'heading5'
  | 'heading6'
  | 'heading7'
  | 'paragraphLarge'
  | 'paragraphMedium'
  | 'paragraphSmallMedium'
  | 'paragraphSmall';

export type TypographyWeightVariant = 'regular' | 'bold';

export type TypographyAlignVariant = 'left' | 'center' | 'right';